+++ trunk/VersionIFORT/int2G.f
@@ -25,7 +25,7 @@ c-----------------------------------------------------------------
       subroutine int2G(NORM,natom,r,Nucd,M,Md,ncontd,nshelld,cd,ad,
      >           cf,f)
 c
-      implicit real*8 (a-h,o-z)
+      implicit real*16 (a-h,o-z)
       logical NORM
       INCLUDE 'param'
       parameter(pi32=5.56832799683170698D0,pi=3.14159265358979312D0,
